Refactors WMSFeatureInfoWidget async usage pattern to add support for geotiff ktile layer.
