+++ web_external/templates/widgets/featureInfoContent.pug
