+++ web_external/templates/widgets/featureInfoWidget.pug
